requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3